build libs with new resampling method

+++ faimed3d/augment.py
@@ -374,9 +374,9 @@ def grid_tfms(t: (TensorDicom3D,TensorMask3D), func, mode):
     meshx, meshy = func(h,w)
     grid = torch.stack((meshx, meshy),2)
     grid = torch.stack((grid.to(dev),)*t.shape[0],0)
-    out = F.grid_sample(t, grid, align_corners=True, mode = mode, padding_mode='reflection')
+    out = F.grid_sample(t.float(), grid, align_corners=True, mode = mode, padding_mode='reflection')
     out = out.permute(0,1,3,2).contiguous()
-    if typ == TensorMask3D: out=out.round()
+    if typ == TensorMask3D: out=out.round().long()
     return retain_type(out, typ = typ)
 
 class RandomWarp3D(RandTransform):
